requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2.4
django-localflavor==5.0
django-parler==2.3
django-rosetta==0.10.2
fonttools==4.59.0
idna==3.10
pillow==11.3.0
polib==1.2.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
python-stdnum==2.1
requests==2.32.4
sqlparse==0.5.3
stripe==12.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
